requirements.txt
# requirements.txt

openai
pyyaml
python-dotenv
python-binance
ccxt
pytest
flake8
websockets
aiohttp
optuna
torch
fastapi
uvicorn
gym
stable-baselines3
shimmy>=2.0
fastapi
uvicorn
pytest-asyncio>=0.20.0
